reshuffled the inspector module
